requirements.dev.txt
sphinx
numpy
sklearn
scikit-learn
matplotlib
tensorflow-cpu
seaborn
pydot
graphviz
pylint
pandas
scipy
pytest
pyyaml
mlflow
streamlit
scikit-image
opencv-python
tqdm
sphinx-rtd-theme
torch
torchvision
pillow==8.2.0
xgboost
jupyter






